airflow-docker/dags/requirements.txt
apache-airflow==2.5.0
apache-airflow-providers-http==2.5.0
apache-airflow-providers-kafka==2.5.0
pandas
requests
reddit_pipeline
